requirements.txt
opencv-python
pillow
ultralytics
insightface
onnxruntime
